server/requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
distro==1.9.0
exceptiongroup==1.3.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Executor==1.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
groqcloud==0.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.6
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
